refactor(db): move single tag query to repository (DB-REPOSITORY-PILOT-NEXT-D1-B)
- 新增 tagsRepository.getOwner() 方法
- 新增 tagsRepository.getById() 方法
- 新增 tagsRepository.getByIdWithCount() 方法
- 在 types.ts 中添加 Tag 类型
- 修改 routes/tags.ts 调用 Repository
